+++ cmd/zetae2e/local/local.go
@@ -344,8 +344,8 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 
 	if !skipRegular {
 		// start the EVM tests
-		//startEVMTests(&eg, conf, deployerRunner, verbose)
-		//startBitcoinTests(&eg, conf, deployerRunner, verbose, light, skipBitcoinSetup)
+		startEVMTests(&eg, conf, deployerRunner, verbose)
+		startBitcoinTests(&eg, conf, deployerRunner, verbose, light, skipBitcoinSetup)
 	}
 	if !skipPrecompiles {
 		precompiledContractTests := []string{
@@ -526,32 +526,32 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 	}
 
 	if testLegacy {
-		//eg.Go(legacyERC20TestRoutine(conf, deployerRunner, verbose,
-		//	e2etests.TestLegacyERC20WithdrawName,
-		//	e2etests.TestLegacyMultipleERC20WithdrawsName,
-		//	e2etests.TestLegacyERC20DepositAndCallRefundName))
+		eg.Go(legacyERC20TestRoutine(conf, deployerRunner, verbose,
+			e2etests.TestLegacyERC20WithdrawName,
+			e2etests.TestLegacyMultipleERC20WithdrawsName,
+			e2etests.TestLegacyERC20DepositAndCallRefundName))
 		eg.Go(legacyZETATestRoutine(conf, deployerRunner, verbose,
-			//e2etests.TestLegacyZetaWithdrawName,
-			//e2etests.TestLegacyMessagePassingExternalChainsName,
-			//e2etests.TestLegacyMessagePassingRevertFailExternalChainsName,
-			//e2etests.TestLegacyMessagePassingRevertSuccessExternalChainsName,
-			//e2etests.TestLegacyZetaDepositRestrictedName,
+			e2etests.TestLegacyZetaWithdrawName,
+			e2etests.TestLegacyMessagePassingExternalChainsName,
+			e2etests.TestLegacyMessagePassingRevertFailExternalChainsName,
+			e2etests.TestLegacyMessagePassingRevertSuccessExternalChainsName,
+			e2etests.TestLegacyZetaDepositRestrictedName,
 			e2etests.TestLegacyZetaDepositName,
-			//e2etests.TestLegacyZetaDepositNewAddressName,
+			e2etests.TestLegacyZetaDepositNewAddressName,
 		))
-		//eg.Go(legacyZEVMMPTestRoutine(conf, deployerRunner, verbose,
-		//	e2etests.TestLegacyMessagePassingZEVMToEVMName,
-		//	e2etests.TestLegacyMessagePassingEVMtoZEVMName,
-		//	e2etests.TestLegacyMessagePassingEVMtoZEVMRevertName,
-		//	e2etests.TestLegacyMessagePassingZEVMtoEVMRevertName,
-		//	e2etests.TestLegacyMessagePassingZEVMtoEVMRevertFailName,
-		//	e2etests.TestLegacyMessagePassingEVMtoZEVMRevertFailName,
-		//))
-		//eg.Go(legacyEthereumTestRoutine(conf, deployerRunner, verbose,
-		//	e2etests.TestLegacyEtherWithdrawName,
-		//	e2etests.TestLegacyEtherDepositAndCallName,
-		//	e2etests.TestLegacyEtherDepositAndCallRefundName,
-		//))
+		eg.Go(legacyZEVMMPTestRoutine(conf, deployerRunner, verbose,
+			e2etests.TestLegacyMessagePassingZEVMToEVMName,
+			e2etests.TestLegacyMessagePassingEVMtoZEVMName,
+			e2etests.TestLegacyMessagePassingEVMtoZEVMRevertName,
+			e2etests.TestLegacyMessagePassingZEVMtoEVMRevertName,
+			e2etests.TestLegacyMessagePassingZEVMtoEVMRevertFailName,
+			e2etests.TestLegacyMessagePassingEVMtoZEVMRevertFailName,
+		))
+		eg.Go(legacyEthereumTestRoutine(conf, deployerRunner, verbose,
+			e2etests.TestLegacyEtherWithdrawName,
+			e2etests.TestLegacyEtherDepositAndCallName,
+			e2etests.TestLegacyEtherDepositAndCallRefundName,
+		))
 	}
 
 	// while tests are executed, monitor blocks in parallel to check if system txs are on top and they have biggest priority
@@ -578,9 +578,9 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 	noError(deployerRunner.WithdrawEmissions())
 	// Migrate funds from old connector to the new one if there are any funds present
 	if utils.MinimumVersionCheck("v30.0.0", deployerRunner.GetZetacoredVersion()) && testV2ConnectorMigration {
-		fn := V2ConnectorMigrationsTestRoutine(conf, deployerRunner, verbose, []string{
+		fn := V2ZetaTestRoutine(conf, deployerRunner, verbose, []string{
 			e2etests.TestMigrateConnectorFundsName,
-			e2etests.TestV2ZetaDepositName,
+			//e2etests.TestV2ZetaDepositName,
 			e2etests.TestLegacyZetaDepositName,
 			e2etests.TestLegacyMessagePassingExternalChainsName,
 		}...)
@@ -625,8 +625,6 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 
 	// Verify that the balance of restricted address is zero
 	deployerRunner.EnsureZeroBalanceOnRestrictedAddressZEVM()
-
-	//deployerRunner.MigrateConnector()
 
 	if !deployerRunner.IsRunningUpgrade() {
 		// Verify that there are no stale ballots left over after tests complete