+++ zetaclient/chains/evm/observer/v2_inbound.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"context"
 	"encoding/hex"
-	"fmt"
 	"sort"
 
 	"cosmossdk.io/errors"
@@ -175,16 +174,15 @@ func (ob *Observer) parseAndValidateDepositEvents(
 // newDepositInboundVote creates a MsgVoteInbound message for a Deposit event
 func (ob *Observer) newDepositInboundVote(event *gatewayevm.GatewayEVMDeposited) types.MsgVoteInbound {
 
-	fmt.Println("newDepositInboundVote called with event:", event.Raw.TxHash.Hex())
-	fmt.Println("Sender:", event.Sender.Hex())
-	fmt.Println("Receiver:", event.Receiver.Hex())
-	fmt.Println("Amount:", event.Amount.String())
-	fmt.Println("Asset:", event.Asset.Hex())
-
+	coinType := coin.CoinType_ERC20
 	// if event.Asset is zero, it's a native token
-	coinType := coin.CoinType_Zeta
-	if crypto.IsEmptyAddress(event.Asset) {
+	switch {
+	case crypto.IsEmptyAddress(event.Asset):
 		coinType = coin.CoinType_Gas
+	case event.Asset == ethcommon.HexToAddress(ob.ChainParams().ZetaTokenContractAddress):
+		coinType = coin.CoinType_Zeta
+	default:
+		coinType = coin.CoinType_ERC20
 	}
 
 	// to maintain compatibility with previous gateway version, deposit event with a non-empty payload is considered as a call